+++ NEWS.md
@@ -1,5 +1,7 @@
 # Packrat 0.4.4 (Unreleased)
 
+- Packrat now infers whether a project implicitly depends on Shiny.
+
 - Packrat now properly infers whether a package is from a CRAN-like repository,
   by checking if the `source` field maps to one of the repository names in
   `getOption('repos')`. (#185)