+++ NEWS.md
@@ -1,7 +1,12 @@
 # Packrat 0.4.4 (Unreleased)
 
-- Packrat issues a warning if it was unable to infer the source of a particular
-  package on initialization and instead uses the latest CRAN version.
+- The cache directory layout has been modified to ensure help (`?`) calls
+  succeed. This is a breaking change with older versions of Packrat, and so
+  newer versions of Packrat will use a new cache folder. (#194)
+  
+- Packrat issues a warning on `packrat::init()` if it was unable to infer the
+  source of a particular package on initialization and instead uses the latest
+  CRAN version.
 
 - Packrat now infers whether a project implicitly depends on Shiny.